pyproject.toml
[tool.poetry]
name = "jasonbotv4"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
python-dotenv = "^0.19.0"
replit = "^3.2.4"
Flask = "^2.0.1"
neural = "^0.1.0"
neuralintents = "^0.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
